feature(SetRadioComponentLoadOptionForm.cs): класс RadioComponentReadyToLoadEventArgs перемещен в файл RadioComponentReadyToLoadEventArgs.cs
