+++ Lab1View/Lab1View.csproj
@@ -62,6 +62,7 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RadioComponentCreatedEventArgs.cs" />
+    <Compile Include="RadioComponentReadyToLoadEventArgs.cs" />
     <Compile Include="SearchFinishedEventArgs.cs" />
     <Compile Include="SearchRadioComponentForm.cs">
       <SubType>Form</SubType>
