+++ Lab1View/RadioComponentReadyToLoadEventArgs.cs
@@ -1,0 +1,29 @@
+﻿using System;
+
+
+namespace Lab1View
+{
+	/// <summary>
+	/// Класс данных события подтверждения выбора
+	/// параметра загрузки радиокомпонента
+	/// </summary>
+	public class RadioComponentReadyToLoadEventArgs : EventArgs
+	{
+		/// <summary>
+		/// Создает объект класса
+		/// <see cref="RadioComponentReadyToLoadEventArgs"/>
+		/// </summary>
+		/// <param name="radioComponentLoadOption">
+		/// Параметр сохранения радиокомпонентов</param>
+		public RadioComponentReadyToLoadEventArgs(
+			RadioComponentLoadOption radioComponentLoadOption)
+		{
+			RadioComponentLoadOption = radioComponentLoadOption;
+		}
+
+		/// <summary>
+		/// Позволяет получить параметр загрузки радиокомпонентов
+		/// </summary>
+		public RadioComponentLoadOption RadioComponentLoadOption { get; }
+	}
+}
