+++ Lab1View/SetRadioComponentLoadOptionForm.cs
@@ -70,28 +70,4 @@ namespace Lab1View
 		AddToEnd,
 		ReplaceAll
 	}
-
-	/// <summary>
-	/// Класс данных события подтверждения выбора
-	/// параметра загрузки радиокомпонента
-	/// </summary>
-	public class RadioComponentReadyToLoadEventArgs : EventArgs
-	{
-		/// <summary>
-		/// Создает объект класса
-		/// <see cref="RadioComponentReadyToLoadEventArgs"/>
-		/// </summary>
-		/// <param name="radioComponentLoadOption">
-		/// Параметр сохранения радиокомпонентов</param>
-		public RadioComponentReadyToLoadEventArgs(
-			RadioComponentLoadOption radioComponentLoadOption)
-		{
-			RadioComponentLoadOption = radioComponentLoadOption;
-		}
-
-		/// <summary>
-		/// Позволяет получить параметр загрузки радиокомпонентов
-		/// </summary>
-		public RadioComponentLoadOption RadioComponentLoadOption { get; }
-	}
 }
